Use Package in ShrinkwrapChecker
