+++ lib/dependency-checker.js
@@ -19,7 +19,6 @@ function EmberCLIDependencyChecker(project, reporter) {
 }
 
 EmberCLIDependencyChecker.prototype.checkDependencies = function() {
-
   if (alreadyChecked || process.env.SKIP_DEPENDENCY_CHECKER) {
     return;
   }
