+++ lib/package.js
@@ -1,10 +1,15 @@
 'use strict';
 
-function Package(name, versionSpecified, versionInstalled) {
+function Package(name, versionSpecified, versionInstalled, strictCheck) {
   this.name = name;
   this.versionSpecified = versionSpecified;
   this.versionInstalled = versionInstalled;
-  this.needsUpdate = this.updateRequired();
+
+  if (strictCheck) {
+    this.needsUpdate = this.versionSpecified !== this.versionInstalled;
+  } else {
+    this.needsUpdate = this.updateRequired();
+  }
 }
 
 Package.prototype = Object.create({});
