+++ lib/shrinkwrap-checker.js
@@ -1,40 +1,22 @@
 'use strict';
 
+var readPackageJSON = require('./utils/read-package-json');
+var Package = require('./package');
 var path = require('path');
-var fs = require('fs');
 
-function readFile(path){
-  if (fs.existsSync(path)) {
-    return fs.readFileSync(path);
-  }
-}
-
-var ShrinkWrapChecker = function(root, name, version){
+var ShrinkWrapChecker = function(root, name, versionSpecified){
   this.root = root;
   this.name = name;
-  this.version = version;
+  this.versionSpecified = versionSpecified;
 };
 
 ShrinkWrapChecker.prototype.check = function() {
-  var packageJSON = this.readPackageJSON() || {};
+  var packageJSON = readPackageJSON(this.root) || {};
+  var versionInstalled = packageJSON.version,
+      strictVersionCheck = true;
 
-  var packageData = {
-    name: this.name,
-    versionSpecified: this.version,
-    versionInstalled: packageJSON.version
-  };
-
-  packageData.needsUpdate = packageData.versionSpecified !== packageData.versionInstalled;
-  return packageData;
-};
-
-ShrinkWrapChecker.prototype.readPackageJSON = function() {
-  var filePath = path.join(this.root, 'package.json');
-  try {
-    return JSON.parse(readFile(filePath));
-  } catch (e) {
-    return null;
-  }
+  return new Package(
+    this.name, this.versionSpecified, versionInstalled, strictVersionCheck);
 };
 
 ShrinkWrapChecker.checkDependencies = function(root, shrinkWrapJSON) {
